requirements.txt
tigeropen
keyboard
requests
openpyxl
xlsxwriter
findatapy
pytz
moomoo-api
pandas
pandas_market_calendars
yfinance
mplcursors
memory_profiler
alpaca-py
yahoo_fin
discord
supabase
jproperties
ujson